pyproject.toml
[project]
name = "audiobooks"
version = "0.1.0"
description = "Audiobook processing pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ebooklib>=0.20",
    "kokoro>=0.9.4",
    "soundfile>=0.13.1",
    "ipython>=9.7.0",
    "ipykernel>=7.1.0",
    "numpy>=2.3.5",
    "lxml>=6.0.2",
    "ipywidgets>=8.1.8",
    "opencv-python>=4.11.0.86",
    "pathvalidate>=3.3.1",
    "subprocess-run>=0.0.8",
    "pyyaml>=6.0.3",
    "wordfreq>=3.1.1",
    "dpath>=2.2.0",
    "types-lxml>=2025.11.25",
    "pip>=25.3",
    "scipy>=1.16.3",
    "resampy>=0.4.3",
]

[project.optional-dependencies]
cpu = [
  "torch==2.9.1",
]

cu128 = [
  "torch==2.9.1+cu128",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "line-profiler>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
